// src/vector-store-files/execution/vector_db/types.ts
/**
 * Copyright 2024 IBM Corp.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { z } from 'zod';

export const VectorDbDocumentSchema = z.object({
  vector: z.number().array(),
  text: z.string(),
  vectorStoreFileId: z.string(),
  source: z
    .object({
      file: z.object({ id: z.string(), name: z.string() }).optional(),
      url: z.string().optional()
    })
    .optional()
});

export type VectorDbDocument = z.infer<typeof VectorDbDocumentSchema>;

export interface DocumentStats {
  byteUsage: number;
  totalCreated: number;
}